% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma1n2_mcov2.R
\name{gamma1n2.mcov2}
\alias{gamma1n2.mcov2}
\title{Internal function for computing cross-covariance of wombling measures for Matern \eqn{\nu=5/2}}
\usage{
gamma1n2.mcov2(coords, t, u, s0, phi)
}
\arguments{
\item{coords}{coordinates}

\item{t}{value of t}

\item{u}{vector of u}

\item{s0}{starting point on curve \eqn{s_0}}

\item{phi}{posterior sample of \eqn{\phi}}
}
\description{
Internal function for computing cross-covariance of wombling measures for Matern \eqn{\nu=5/2}
}
\keyword{gamma1n2.mcov2}
