% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_int.R
\name{gamma_int}
\alias{gamma_int}
\title{Incomplete Gamma Function}
\usage{
gamma_int(x, a, b)
}
\arguments{
\item{x}{gamma quantile}

\item{a}{shape parameter}

\item{b}{scale parameter}
}
\description{
internal only
}
\keyword{gamma_int}
