% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussian}
\alias{gaussian}
\title{Squared Exponential Covariance with \eqn{\nu = \infty}}
\usage{
gaussian(dists, phi, sigma2, tau2)
}
\arguments{
\item{dists}{distance matrix}

\item{phi}{spatial range}

\item{sigma2}{spatial variance}

\item{tau2}{nugget variance}
}
\description{
Computes the Matern covariance matrix with fractal parameter \eqn{\nu = \infty}.
}
\keyword{materncov2}
