% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients_matern1.R
\name{gradients_matern1}
\alias{gradients_matern1}
\title{Posterior samples of rates of change for Matern with \eqn{\nu=3/2}}
\usage{
gradients_matern1(dists.1, dists.2, dists.3, z, phi, sigma2)
}
\arguments{
\item{dists.1}{distance matrix generated from coordinates}

\item{dists.2}{distance of grid from coordinates}

\item{dists.3}{delta = coordinate - grid}

\item{z}{posterior samples of \eqn{Z(s)}}

\item{phi}{posterior samples of \eqn{\phi}}

\item{sigma2}{posterior samples of \eqn{\sigma^2}}
}
\description{
For internal use only.
}
\keyword{gradients_matern1}
