% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materncov1.R
\name{materncov1}
\alias{materncov1}
\title{Matern Covariance with \eqn{\nu = 3/2}}
\usage{
materncov1(dists, phi, sigma2, tau2)
}
\arguments{
\item{dists}{distance matrix}

\item{phi}{spatial range}

\item{sigma2}{spatial variance}

\item{tau2}{nugget variance}
}
\description{
Computes the Matern covariance matrix with fractal parameter \eqn{\nu = 3/2}.
}
\keyword{materncov1}
