% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnorm_nimble.R
\name{pnorm_nimble}
\alias{pnorm_nimble}
\title{Computes the Cumulative Distribution Function for the standard Gaussian
probability distribution}
\usage{
pnorm_nimble(x)
}
\arguments{
\item{x}{standard Gaussian quantile}
}
\description{
For internal use only.
}
\examples{
\dontrun{
#####################
# Internal use only #
#####################
# Example usage inside of nimblewomble::wombling_gaussian()
pnorm_nimble(sqrt(2) * phi[i] * tvec[j]) - 1)
}
}
\author{
Aritra Halder <aritra.halder@drexel.edu>, \cr
Sudipto Banerjee <sudipto@ucla.edu>
}
\keyword{pnorm_nimble}
