% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{significance}
\alias{significance}
\title{Internal Function for significance}
\usage{
significance(data_frame = NULL)
}
\arguments{
\item{data_frame}{matrix consisting of median, lower and upper CI}
}
\description{
Internal Function for significance
}
\keyword{sig}
