% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_plot.R
\name{sp_ggplot}
\alias{sp_ggplot}
\title{Spatial Plot Function}
\usage{
sp_ggplot(
  data_frame = NULL,
  sp = FALSE,
  shape = NULL,
  legend.key.height = 0.7,
  legend.key.width = 0.4,
  text.size = 10,
  palette = "Spectral"
)
}
\arguments{
\item{data_frame}{data frame consisting of coordinates and data}

\item{sp}{logical parameter indicating whether to make a spatial plot}

\item{shape}{if sp = TRUE shape file should be provided (should be an sf object)}

\item{legend.key.height}{height of legend (defaults to .7)}

\item{legend.key.width}{width of legend (defaults to .4)}

\item{text.size}{size of legend text (defaults to 10)}

\item{palette}{(optional) color palette}
}
\description{
Spatial Plot Function
}
\keyword{sp_plot}
