% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprates.R
\name{sprates}
\alias{sprates}
\title{Posterior samples for rates of change}
\usage{
sprates(
  coords = NULL,
  grid = NULL,
  model = NULL,
  kernel = c("matern1", "matern2", "gaussian")
)
}
\arguments{
\item{coords}{coordinates}

\item{grid}{grid for sampling the rates of change}

\item{model}{posterior samples of \eqn{Z(s)}, \eqn{\phi}, \eqn{\sigma^2}}

\item{kernel}{choice of kernel; must be one of "matern1", "matern2", "gaussian"}
}
\description{
Posterior samples for rates of change
}
\keyword{sprates}
