% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbeta_gaussian.R
\name{zbeta_gaussian}
\alias{zbeta_gaussian}
\title{Posterior samples of spatial effects and intercept for the squared exponential kernel}
\usage{
zbeta_gaussian(y, dists, phi, sigma2, tau2)
}
\arguments{
\item{y}{response}

\item{dists}{distance matrix derived from coordinates}

\item{phi}{posterior samples of \eqn{\phi}}

\item{sigma2}{posterior samples of \eqn{\sigma^2}}

\item{tau2}{posterior samples of \eqn{\tau^2}}
}
\description{
For internal use only.
}
\keyword{zbeta_gaussian}
