% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbeta_samples.R
\name{zbeta_samples}
\alias{zbeta_samples}
\title{Posterior samples of spatial effects and intercept for Matern with \eqn{nu=3/2}}
\usage{
zbeta_samples(
  coords = NULL,
  y = NULL,
  X = NULL,
  model = NULL,
  kernel = c("matern1", "matern2", "gaussian")
)
}
\arguments{
\item{coords}{coordinates}

\item{y}{response}

\item{X}{covariates (supply as a matrix without intercept)}

\item{model}{matrix of posterior samples of \eqn{\phi}, \eqn{\sigma^2} and \eqn{\tau^2}}

\item{kernel}{choice of kernel; must be one of "matern1", "matern2", "gaussian"}
}
\description{
For internal use only.
}
\keyword{zbeta_samples}
