% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zXbeta.R
\name{zXbeta}
\alias{zXbeta}
\title{Posterior samples of spatial effects and intercept for all kernels in the
presence of covariates}
\usage{
zXbeta(y, X, beta)
}
\arguments{
\item{y}{response}

\item{X}{covariates (supply as a matrix without intercept)}

\item{beta}{posterior samples of \eqn{\beta} (supply as a matrix)}
}
\description{
For internal use only.
}
\examples{
\dontrun{
#####################
# Internal use only #
#####################
# Example usage inside of nimblewomble::zbeta_samples(...)
zXb = nimble::compileNimble(zXbeta)
zb.samples = zXb(y = y, X = X, beta = beta)
}
}
\author{
Aritra Halder <aritra.halder@drexel.edu>, \cr
Sudipto Banerjee <sudipto@ucla.edu>
}
\keyword{zXbeta}
