% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma1_mcov1.R
\name{gamma1.mcov1}
\alias{gamma1.mcov1}
\title{Internal function for computing cross-covariance of wombling measures for Matern \eqn{\nu=3/2}}
\usage{
gamma1.mcov1(coords, t, u, s0, phi)
}
\arguments{
\item{coords}{coordinates}

\item{t}{value of t}

\item{u}{vector of u}

\item{s0}{starting point on curve \eqn{s_0}}

\item{phi}{posterior sample of \eqn{\phi}}
}
\description{
Internal function for computing cross-covariance of wombling measures for Matern \eqn{\nu=3/2}
}
\keyword{gamma1.mcov1}
