% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussian}
\alias{gaussian}
\title{Squared Exponential Covariance kernel}
\usage{
gaussian(dists, phi, sigma2, tau2)
}
\arguments{
\item{dists}{distance matrix}

\item{phi}{spatial range}

\item{sigma2}{spatial variance}

\item{tau2}{nugget variance}
}
\description{
Computes the Matern covariance matrix with fractal parameter
\eqn{\nu \to \infty}. Has the option to compute \eqn{\Sigma_{d\times d} + \tau^2 I_d}.
}
\examples{
\dontrun{
#####################
# Internal use only #
#####################
# Used across multiple functions
# Example usage
gaussian(dists = dists[1:ncoords, 1:ncoords], phi = phi[i], sigma2 = 1, tau2 = 0)
}
}
\author{
Aritra Halder <aritra.halder@drexel.edu>, \cr
Sudipto Banerjee <sudipto@ucla.edu>
}
\keyword{materncov2}
