% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_fit.R
\name{gp_fit}
\alias{gp_fit}
\title{Fit a Gaussian process}
\usage{
gp_fit(
  coords = NULL,
  y = NULL,
  kernel = c("matern1", "matern2", "gaussian"),
  niter = NULL,
  nburn = NULL
)
}
\arguments{
\item{coords}{spatial coordinats (supply as a matrix)}

\item{y}{response}

\item{kernel}{choice of kernel; must be one of "matern1", "matern2", "gaussian"}

\item{niter}{number of iterations}

\item{nburn}{burn-in}
}
\description{
Fits a Gaussian process with the choice of three kernels. Uses `nimble` to generate posterior samples.
}
\keyword{gp_fit}
