% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wombling_matern1.R
\name{wombling_matern1}
\alias{wombling_matern1}
\title{Posterior samples for wombling measures from Matern with \eqn{\nu=3/2}}
\usage{
wombling_matern1(coords, curve, dists, tvec, umat, z, phi, sigma2)
}
\arguments{
\item{coords}{coordinates}

\item{curve}{curve coordinates}

\item{dists}{distance matrix}

\item{tvec}{vector of t's}

\item{umat}{matrix of u's}

\item{z}{posterior samples of \eqn{Z(s)}}

\item{phi}{posterior samples of \eqn{\phi}}

\item{sigma2}{posterior samples of \eqn{\sigma^2}}
}
\description{
Posterior samples for wombling measures from Matern with \eqn{\nu=3/2}
}
\keyword{wombling_matern1}
