% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zXbeta.R
\name{zXbeta}
\alias{zXbeta}
\title{Posterior samples of spatial effects and intercept for Matern with \eqn{nu=3/2}}
\usage{
zXbeta(y, X, beta)
}
\arguments{
\item{y}{response}

\item{X}{covariates (supply as a matrix without intercept)}

\item{beta}{posterior samples of \eqn{\beta} (supply as a matrix)}
}
\description{
For internal use only.
}
\keyword{zXbeta}
